feature/17-Added Authentication-checks-and-route-protection

+++ client/src/components/UserDetails.js
@@ -1,13 +1,21 @@
 import { useDetailsContext } from '../hooks/useDetailsContext'
 import formatDistanceToNow from 'date-fns/formatDistanceToNow';
 import format from 'date-fns/format';
+import { useAuthContext } from '../hooks/useAuthContext';
 
 const UserDetails = ({ detail }) => {
     const { dispatch } = useDetailsContext()
+    const { user } = useAuthContext()
 
     const handleClick = async () => {
+        // if check for logged in user
+        if(!user) {
+            return
+        }
         const response = await fetch('api/details/' + detail._id, {
-            method: 'DELETE'
+            method: 'DELETE',
+            // add auth check for valid token 
+            headers: {'Authorization': `Bearer ${user.token}`}
         });
     
         const json = await response.json()
